backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
python-dotenv>=1.0
anthropic>=0.40.0
requests>=2.31.0